Put box-blur and median-blur in the gegl workshop folder
so they show up in the gegl operations menu in GIMP.

I'm not sure how to add these operations to the GIMP UI using
the new way to show GEGL ops in GIMP. So this is a quick fix.

In default GIMP box-blur has been removed from the UI and
median-blur is still in the gegl workshop.