Stretch Contrast options to only adjust white or black point
Sometimes it's very convenient to only normalize the white point
or the black point.

Technically this operation is "Normalize", not "Stretch-Contrast",
when working with potentially out of display range channel values.

But to have the operation renamed in GIMP, probably something in
GIMP also needs to be changed, so I left the operation name alone.

I did change the category to "color", in place of "color:enhance",
as normalizing an HDR image is seldom considered an enhancement in
and of itself.